third_party/blink/web_tests/fast/forms/label/label-click-expected.txt

 Label
PASS successfullyParsed is true

TEST COMPLETE

PASS checkbox.checked is false
PASS checkbox.checked is true
PASS checkbox.checked is false
PASS checkbox.checked is true
PASS checkbox.checked is false
PASS checkbox.checked is true
PASS checkbox.checked is false
